+++ src/routes/config.ts
@@ -11,6 +11,9 @@ export class ConfigRouter {
 
     this.router.get('/', users.ensureAdmin, this.getConfig);
     this.router.post('/', users.ensureAdmin, this.updateConfig);
+    this.router.get('/backups', users.ensureAdmin, this.listConfigBackups);
+    this.router.get('/backups/:backupId(\\d+)', users.ensureAdmin, this.getConfigBackup);
+    this.router.delete('/backups', users.ensureAdmin, this.deleteAllConfigBackups);
   }
 
   getConfig(req: Request, res: Response, next: NextFunction) {
@@ -21,6 +24,30 @@ export class ConfigRouter {
     return hb.updateConfig(req.body)
       .then((config) => {
         res.json(config);
+      })
+      .catch(next);
+  }
+
+  listConfigBackups(req: Request, res: Response, next: NextFunction) {
+    return hb.listConfigBackups()
+      .then((data) => {
+        res.json(data);
+      })
+      .catch(next);
+  }
+
+  getConfigBackup(req: Request, res: Response, next: NextFunction) {
+    return hb.getConfigBackup(req.params.backupId)
+      .then((backupConfig) => {
+        res.send(backupConfig);
+      })
+      .catch(next);
+  }
+
+  deleteAllConfigBackups(req: Request, res: Response, next: NextFunction) {
+    return hb.deleteAllConfigBackups()
+      .then((backupConfig) => {
+        res.json({ok: true});
       })
       .catch(next);
   }