+++ ui/src/app/plugins/plugins.manage.component.ts
@@ -136,13 +136,6 @@ export class PluginsManageComponent implements OnInit {
   public onRestartHomebridgeClick() {
     this.$state.go('restart');
     this.activeModal.close();
-
-    // make sure the page is refreshed after restart
-    setInterval(() => {
-      if (this.$state.current.name !== 'restart') {
-        window.location.reload(true);
-      }
-    }, 1000);
   }